pyproject.toml
[project]
name = "document-checker-exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pdfplumber>=0.11.9",
    "python-docx>=1.2.0",
]
